Keep quad tree balanced on update
Fixes a small bug, when updating a QTObject boundaries inside the tree.
On update, if the object still fits in node, we should also check
if it fits in a child, because maybe we previously climbed up to this
node from a child one -> AKA the object traversed from
child[NE] -> child[SE], but because it's boundaries are inside both, it
was attached to the parent. This fix the tree balanced.
Also fixes the cameraGetAABB (gets world coordinates the camera sees).